@@ -25,7 +25,7 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SharedIndicatorComponent } from '@shared/components/shared-indicator/shared-indicator.component';
-
+import { DateService } from '@core/services/date.service';
 
 describe('DashboardDetailComponent', () => {
   let component: DashboardDetailComponent;
