@@ -1,4 +1,4 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Component } from '@angular/core';
 import { MetricsService } from '@features/metrics/services/metrics.service';
 import { Subscription } from 'rxjs';
 
@@ -7,16 +7,10 @@ import { Subscription } from 'rxjs';
   templateUrl: './metrics.component.html',
   styleUrls: ['./metrics.component.scss']
 })
-export class MetricsComponent implements OnInit {
+export class MetricsComponent {
   subscription: Subscription = new Subscription();
   constructor(
     private metricsService: MetricsService
   ) { }
-
-  ngOnInit() {
-
-
-  }
-
 
 }
