@@ -1,14 +1,13 @@
-import { Injectable, OnDestroy } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { SquacApiService } from '@core/services/squacapi.service';
-import { Observable, Subject, Subscription } from 'rxjs';
+import { Observable, Subject } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { Alert, AlertAdapter } from '../models/alert';
-import { Trigger } from '../models/trigger';
 
 @Injectable({
   providedIn: 'root'
 })
-export class AlertsService implements OnDestroy{
+export class AlertsService {
   localAlerts: {[monitorId: number]: Alert[]} = {};
   alerts: Subject<Alert[]> = new Subject();
   refreshTimeout;
@@ -20,9 +19,6 @@ export class AlertsService implements OnDestroy{
     private squacApi: SquacApiService,
     private alertAdapter: AlertAdapter
   ) {}
-
-  ngOnDestroy(): void {
-  }
 
   getAlerts(): Observable<Alert[]>{
     return this.squacApi.get(this.url).pipe(
