@@ -5,13 +5,10 @@ import { Component, Input, OnInit } from '@angular/core';
   templateUrl: './shared-indicator.component.html',
   styleUrls: ['./shared-indicator.component.scss']
 })
-export class SharedIndicatorComponent implements OnInit {
+export class SharedIndicatorComponent {
   @Input() shareOrg: boolean;
   @Input() shareAll: boolean;
   @Input() resource: string;
   constructor() { }
-
-  ngOnInit(): void {
-  }
 
 }
